+++ cli/src/subcommands/mod.rs
@@ -1,2 +1,3 @@
 pub mod new;
 pub mod package;
+pub mod run;