+++ cli/src/subcommands/package.rs
@@ -3,6 +3,7 @@ use std::fs;
 use std::path::Path;
 use rust_embed::RustEmbed;
 use serde_json::{Value, Map};
+use crate::manifest::read_manifest;
 
 #[derive(StructOpt, Debug)]
 #[structopt(name = "package")]
@@ -46,37 +47,17 @@ pub fn package_command(opts: PackageOpts) -> Result<(), String> {
   let manifest_path = opts.manifest_path.unwrap_or("manifest.json".to_string());
   let manifest_path = Path::new(&manifest_path);
   let manifest_directory = manifest_path.parent().unwrap();
+  
+  let mut manifest_dict = read_manifest(manifest_path)?;
 
-  let manifest_data = fs::read(manifest_path)
-    .map_err(|_| format!("Error reading manifest file. Are you in the right directory?"))?;
-  
-  let mut v: Value = serde_json::from_slice(&manifest_data)
-    .map_err(|e| format!("Error parsing manifest: {}", e))?;
-  
   let dist_directory = manifest_directory.join(Path::new("dist"));
   
   fs::create_dir_all(&dist_directory);
   
-  let manifest_dict = v.as_object_mut().ok_or("Invalid manifest. Expected a JSON object at the root.")?;
+  copy_and_update_path(&mut manifest_dict, manifest_directory, "program", "program.wasm")?;
+  copy_and_update_path(&mut manifest_dict, manifest_directory, "palette", "palette.png")?;
   
-  let manifest_version = manifest_dict.get("manifestVersion")
-    .ok_or("Missing `manifestVersion` key")?
-    .as_i64()
-    .ok_or("Expected `manifestVersion` to be an integer")?;
-  
-  if manifest_version > MAX_SUPPORTED_MANIFEST_VERSION {
-    let cli_path = std::env::current_exe().unwrap();
-    let cli_name = cli_path.file_name().unwrap().to_str().unwrap();
-    eprintln!(
-      "WARNING: Manifest version {} not supported, things may not behave as expected. Maximum version supported is {}. Do you need to upgrade {}?",
-      manifest_version, MAX_SUPPORTED_MANIFEST_VERSION, cli_name
-    )
-  }
-  
-  copy_and_update_path(manifest_dict, manifest_directory, "program", "program.wasm")?;
-  copy_and_update_path(manifest_dict, manifest_directory, "palette", "palette.png")?;
-  
-  let fonts_dict = manifest_dict
+  let fonts_dict = &mut manifest_dict
     .get_mut("fonts")
     .ok_or("Manifest missing `fonts` key")?
     .as_object_mut()
@@ -91,7 +72,7 @@ pub fn package_command(opts: PackageOpts) -> Result<(), String> {
     fs::write(destination_path, file.data);
   }
   
-  let manifest_data = serde_json::to_vec_pretty(manifest_dict).unwrap();
+  let manifest_data = serde_json::to_vec_pretty(&mut manifest_dict).unwrap();
   fs::write((&dist_directory).join("manifest.json"), manifest_data);
   
   Ok(())